# src/v3/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova API — v3",
    "version": "3.0.0",
    "description": "API v3 - Condomínio Nova Residence"
  },
...
  "components": {
    "schemas": {
      "ResponseMeta": {
        "type": "object",
        "description": "Metadado transversal presente em toda resposta, dentro ou fora do envelope de sucesso/erro.",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Instante em que a resposta foi gerada.",
            "example": "2026-09-28T03:44:40.342Z"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "description": "Identificador da requisição, para correlação em logs e suporte.",
            "example": "req-1"
          },
          "version": {
            "type": "string",
            "enum": ["v3"],
            "description": "Versão da API que gerou a resposta."
          },
          "pagination": {
            "type": "object",
            "description": "Presente somente em endpoints que retornam uma coleção paginada.",
            "properties": {
              "limit": { "type": "integer", "minimum": 1, "example": 25 },
              "offset": { "type": "integer", "minimum": 0, "example": 0 },
              "total": { "type": "integer", "minimum": 0, "example": 137 }
            },
            "required": ["limit", "offset", "total"]
          }
        },
        "required": ["timestamp", "requestId", "version"]
      },
      "ProblemDetails": {
        "type": "object",
        "description": "Estrutura padrão de erro da API, com tipo, título, status HTTP e detalhe da ocorrência.",
        "properties": {
          "type": {
            "type": "string",
            "description": "URN fixo identificando o tipo do erro (não é uma URL navegável).",
            "example": "urn:nova-api:validation-error"
          },
          "title": {
            "type": "string",
            "description": "Resumo legível do tipo de erro.",
            "example": "Dados de entrada inválidos"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP correspondente, redundante com o status da resposta.",
            "example": 400
          },
          "detail": {
            "type": "string",
            "nullable": true,
            "description": "Explicação específica desta ocorrência do erro.",
            "example": "Um ou mais campos da requisição são inválidos."
          },
          "instance": {
            "type": "string",
            "nullable": true,
            "description": "Path da requisição que originou o erro.",
            "example": "/v3/api/health"
          },
          "validationErrors": {
            "type": "array",
            "description": "Presente somente quando type = validation-error.",
            "items": {
              "type": "object",
              "properties": {
                "path": { "type": "string", "example": "/name" },
                "message": { "type": "string", "example": "Too small: expected string to have >=3 characters" }
              },
              "required": ["path", "message"]
            }
          }
        },
        "required": ["type", "title", "status", "detail", "instance"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "$ref": "#/components/schemas/ProblemDetails" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "error", "meta"]
      },
...
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "nullable": true },
          "role": {
            "type": "string",
            "nullable": true,
            "enum": ["morador", "porteiro", "sindico", "admin"],
            "description": "Papel da aplicação. null quando a conta ainda não tem papel atribuído."
          }
        },
        "required": ["id", "email", "role"]
      },
      "SessionData": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string", "description": "JWT de acesso. Enviar em Authorization: Bearer <accessToken>." },
          "refreshToken": { "type": "string", "description": "Uso único: cada refresh devolve um novo refresh token." },
          "tokenType": { "type": "string", "enum": ["bearer"] },
          "expiresIn": { "type": "integer", "description": "Segundos até o access token expirar.", "example": 900 },
          "expiresAt": { "type": "integer", "description": "Expiração do access token (Unix epoch, segundos)." },
          "user": { "$ref": "#/components/schemas/AuthUser" }
        },
        "required": ["accessToken", "refreshToken", "tokenType", "expiresIn", "expiresAt", "user"]
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "data": { "$ref": "#/components/schemas/SessionData" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "data", "meta"]
      },
      "AuthUserResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "data": { "$ref": "#/components/schemas/AuthUser" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "data", "meta"]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "data": { "nullable": true, "description": "Sempre null." },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "data", "meta"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado — token ausente, inválido ou expirado.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token obtido em POST /v3/api/auth/login. Expira em 15 minutos; renovar em POST /v3/api/auth/refresh."
      }
    }
  },
  "paths": {
...
    "/v3/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Autentica com e-mail e senha. Limites: 20 tentativas por IP e 5 falhas por e-mail a cada 15 minutos (429 com Retry-After).",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionResponse" } } }
          },
          "400": {
            "description": "Dados de entrada inválidos.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Credenciais inválidas.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "429": {
            "description": "Muitas tentativas.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/v3/api/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Cadastro público",
        "description": "Cria a conta e já devolve a sessão. A conta nasce sem papel e sem acesso (pendente) até um admin vinculá-la ao cadastro do condomínio — ver /residence. Limite: 5 cadastros por IP por hora (429 com Retry-After).",
        "operationId": "authSignup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conta criada e sessão iniciada.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionResponse" } } }
          },
          "400": {
            "description": "Dados de entrada inválidos ou recusados pela política de senha.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "429": { "description": "Muitas tentativas.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": {
            "description": "Já existe uma conta com este e-mail.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/v3/api/auth/users": {
      "get": {
        "tags": ["Auth"],
        "summary": "Listar contas (síndico/admin)",
        "description": "Contas do Supabase Auth, para achar quem se cadastrou e vincular ao cadastro.",
        "operationId": "authListUsers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "perPage", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Contas.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          { "$ref": "#/components/schemas/AuthUser" },
                          {
                            "type": "object",
                            "properties": { "createdAt": { "type": "string", "format": "date-time" } }
                          }
                        ]
                      }
                    },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  },
                  "required": ["success", "data", "meta"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Papel sem permissão para esta rota.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Criar conta (síndico/admin)",
        "description": "Cria uma conta com qualquer papel. Exige síndico ou admin; só admin cria admin. O primeiro admin é promovido direto no banco.",
        "operationId": "authCreateUser",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "role": { "type": "string", "enum": ["morador", "porteiro", "sindico", "admin"] }
                },
                "required": ["email", "password", "role"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conta criada.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthUserResponse" } } }
          },
          "400": {
            "description": "Dados de entrada inválidos ou recusados pela política de senha.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Papel sem permissão para esta rota.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "409": {
            "description": "Já existe uma conta com este e-mail.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/v3/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Renovar sessão",
        "description": "Troca um refresh token por um novo par de tokens.",
        "operationId": "authRefresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "refreshToken": { "type": "string" } },
                "required": ["refreshToken"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionResponse" } } }
          },
          "400": {
            "description": "Dados de entrada inválidos.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Refresh token inválido ou revogado.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "429": {
            "description": "Muitas tentativas.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/v3/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "description": "Revoga a sessão atual (refresh token). O access token segue válido até expirar (máx. 15 min).",
        "operationId": "authLogout",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sessão encerrada.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LogoutResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": {
            "description": "Muitas tentativas.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/v3/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Usuário autenticado",
        "description": "Dados do usuário extraídos do access token.",
        "operationId": "authMe",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Usuário atual.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthUserResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}